requirements.txt
Django
djangorestframework
psycopg2==2.9.4
drf-spectacular==0.24.2
uwsgi>=2.0.19,<2.1
django-rest-durin>=1.0.0,<2.0
django-cors-headers>=3.13.0,<3.14.0
googlemaps
Pillow>=9.2.0,<9.3.0
djangorestframework-csv

pytz
